Flink Delta Source PR 12 - DeltaSource::option(...) value type conversion. (#364)
* PR ColumnsFromDeltaLog - get table schema from Delta Log.


* PR ColumnsFromDeltaLog - get table schema from Delta Log. Using SnapshotVersion in Enumerator factory.


* PR 10.1 ColumnsFromDeltaLog_Tests - extra test for Delta table Schema discovery.


* PR 11 - Partition support using Delta Log Metadata


* PR 10 - Changes after code review.


* PR 10.1 - Added Delta - Flink - Delta type conversion test.


* PR 10 ColumnsFromDeltaLog - Prevent user to set internal options via source builder.


* PR 10 ColumnsFromDeltaLog - Changes after code review


* PR 10.1 Adding tests


* PR 10.1 Adding tests


* PR 11 Partition Support


* PR 12 Option validation - Get BATCH_SIZE for Format builder from Source options. Add validation for option names and values.


* PR 12 Option validation - Get BATCH_SIZE for Format builder from Source options. Add validation for option names and values.


* PR 10.1 Changes after code review.


* PR 10.1 Added tests.


* PR 11 Test fix after merge


* PR 10.1 cleanup.


* PR 11 Fix after merge conflicts from master.


* PR 11 Make RowDataFormat constructor package protected.


* PR 11 Changes after Code review


* PR 12 Fix compilation error after merge from base branch.


* PR 12 - Validation for Inapplicable Option Used + tests.


* PR 12 - Javadocs


* PR 12 - Adding option type safety


* PR 12 - Adding option type safety tests


* PR 11 - Changes after code review.


* PR 11 - Changes after code review.


* PR 11 - Changes after code review.


* PR 11 - Changes after code review.


* PR 11 - Changes after code review.


* PR 12 - test for options


* PR 12 - Type conversion for Timestamp based options and adding tests.


* PR 12 - Change TODO's from PR 12 to PR 12.1


* PR 12 - changes after code review

+++ flink/src/test/java/io/delta/flink/source/DeltaSourceContinuousExecutionITCaseTest.java
@@ -139,7 +139,7 @@ public class DeltaSourceContinuousExecutionITCaseTest extends DeltaSourceITBase 
 
     @ParameterizedTest(name = "{index}: FailoverType = [{0}]")
     @EnumSource(FailoverType.class)
-    // This test updates Delta Table 5 times, so it will take some time to finish. About 1 minute.
+    // This test updates Delta Table 5 times, so it will take some time to finish.
     public void shouldReadDeltaTableFromSnapshotAndUpdatesUsingUserSchema(FailoverType failoverType)
         throws Exception {
 
@@ -270,7 +270,7 @@ public class DeltaSourceContinuousExecutionITCaseTest extends DeltaSourceITBase 
         DeltaSourceConfiguration sourceConfiguration = source.getSourceConfiguration();
         // Main thread waits some time. To stay on a safe side, we wait doubled time of update
         // check interval. So source will have time to check table twice for updates.
-        int testTimeout =
+        long testTimeout =
             sourceConfiguration.getValue(DeltaSourceOptions.UPDATE_CHECK_INTERVAL) * 2;
         List<RowData> results = dataFuture.get(testTimeout, TimeUnit.MILLISECONDS);
